requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
azure-ai-ml==1.23.0
azure-common==1.1.28
azure-core==1.36.0
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-mgmt-core==1.6.0
azure-monitor-opentelemetry-exporter==1.0.0b31
azure-search-documents==11.6.0
azure-storage-blob==12.27.0
azure-storage-file-datalake==12.22.0
azure-storage-file-share==12.23.0
backoff==2.2.1
bandit==1.8.6
bcrypt==5.0.0
black==25.1.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
crewai==0.203.1
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.68.0
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.9.0.post1
fastapi==0.115.0
filelock==3.20.0
fixedint==0.1.6
FlashRank==0.2.9
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.11.3
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-pinecone==0.2.2
langchain-text-splitters==0.3.4
langgraph==0.2.58
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.74
langsmith==0.2.11
litellm==1.74.9
lxml==6.0.2
Mako==1.3.10
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
mcp==1.17.0
mdurl==0.1.2
mlflow==2.19.0
mlflow-skinny==2.19.0
mmh3==5.2.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.2
msrest==0.7.1
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.1
openai==1.109.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
opencensus-ext-logging==0.1.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.5.0
pluggy==1.6.0
portalocker==2.7.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydash==8.0.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==5.1.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-benchmark==5.1.0
pytest-codspeed==4.1.1
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
pywin32==311
PyYAML==6.0.3
rank-bm25==0.2.2
redis==5.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.13.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.38.6
stevedore==5.5.0
strictyaml==1.7.3
sympy==1.14.0
syrupy==4.9.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.4
tomli==2.3.0
tomli_w==1.2.0
torch==2.9.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uv==0.9.3
uvicorn==0.31.1
vcrpy==7.0.0
waitress==3.0.2
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
